fix(driver): getsockopt exit event, send correct (s64)fd

+++ driver/modern_bpf/programs/tail_called/events/syscall_dispatched_events/getsockopt.bpf.c
@@ -61,15 +61,15 @@ int BPF_PROG(getsockopt_x,
 	auxmap__store_s64_param(auxmap, (s64)fd);
 
 	/* Parameter 3: level (type: PT_ENUMFLAGS8) */
-	u8 level = (u8)extract__syscall_argument(regs, 1);
+	int level = extract__syscall_argument(regs, 1);
 	auxmap__store_u8_param(auxmap, sockopt_level_to_scap(level));
 
 	/* Parameter 4: optname (type: PT_ENUMFLAGS8) */
-	u8 optname = (u8)extract__syscall_argument(regs, 2);
+	int optname = extract__syscall_argument(regs, 2);
 	auxmap__store_u8_param(auxmap, sockopt_optname_to_scap(level, optname));
 
 	/* `optval` and `optlen` will be the ones provided by the user if the syscall fails
-	 * otherwise they will be directly extracted from the socket.
+	 * otherwise they will refer to the real socket data since the kernel populated them.
 	 */
 
 	/* Parameter 5: optval (type: PT_DYN) */
